backend/requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.2
sqlalchemy>=2.0.20
python-nmap>=0.7.1
